+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.0.0-beta.7 - 2024-03-29
+
+### Added
+- Add Formie forms as a link type.
+
+### Fixed
+- Fix an error when limiting link types and removing the current link type of a link.
+- Fix Shopify Product link type label.
+- Fix Typed Link migration not including suffixes for element link types.
+- Fix an error where other field content would be cleared when Hyper field content was initialized at the same time.
+- Fix Typed Link migration not including suffixes for element link types.
+
 ## 2.0.0-beta.6 - 2024-03-27
 
 ### Changed
@@ -51,6 +63,18 @@
 ### Changed
 - Now requires PHP `8.2.0+`.
 - Now requires Craft `5.0.0-beta.1+`.
+
+## 1.1.26 - 2024-03-29
+
+### Added
+- Add Formie forms as a link type.
+
+### Fixed
+- Fix an error when limiting link types and removing the current link type of a link.
+- Fix Shopify Product link type label.
+- Fix Typed Link migration not including suffixes for element link types.
+- Fix an error where other field content would be cleared when Hyper field content was initialized at the same time.
+- Fix Typed Link migration not including suffixes for element link types.
 
 ## 1.1.25 - 2024-03-23